+++ sql/vroom/vrp_vroomJobsPlain.sql
@@ -67,7 +67,7 @@ CREATE FUNCTION vrp_vroomJobsPlain(
     TEXT,  -- breaks_time_windows_sql (required)
     TEXT,  -- matrix_sql (required)
 
-    exploration_level SMALLINT DEFAULT 5,
+    exploration_level INTEGER DEFAULT 5,
     timeout INTEGER DEFAULT -1,
 
     OUT seq BIGINT,
@@ -102,7 +102,7 @@ LANGUAGE plpgsql VOLATILE;
 
 -- COMMENTS
 
-COMMENT ON FUNCTION vrp_vroomJobsPlain(TEXT, TEXT, TEXT, TEXT, TEXT, TEXT, SMALLINT, INTEGER)
+COMMENT ON FUNCTION vrp_vroomJobsPlain(TEXT, TEXT, TEXT, TEXT, TEXT, TEXT, INTEGER, INTEGER)
 IS 'vrp_vroomJobsPlain
  - EXPERIMENTAL
  - Parameters:
@@ -120,7 +120,7 @@ IS 'vrp_vroomJobsPlain
    - Matrix SQL with columns:
        start_vid, end_vid, agg_cost
 - Optional parameters
-   - exploration_level := 5::SMALLINT
+   - exploration_level := 5
    - timeout := -1
  - Documentation:
    - ${PROJECT_DOC_LINK}/vrp_vroomJobsPlain.html
